+++ R/app_02_resolve.R
@@ -17,17 +17,16 @@ ghqc_notify_app <- function() {
 
   # error handling before starting app
   check_github_credentials()
-  milestone_list <- get_open_milestone_list_errors()
+  all_milestone_objects <- get_all_non_empty_ghqc_milestone_objects()
+  open_milestone_objects <- get_open_milestone_objects_from_all_milestone_objects(all_milestone_objects)
+  open_milestone_names <- get_milestone_names_from_milestone_objects(open_milestone_objects)
 
   # error if no open ghqc milestones
-  if (length(milestone_list) == 0) {
+  if (length(open_milestone_objects) == 0) {
     error(.le$logger, glue::glue("There were no open ghqc Milestones found in {.le$org}/{.le$repo}. Create ghqc Milestones using `ghqc_assign_app()`"))
     rlang::abort("There were no open ghqc Milestones found.")
   }
 
-  # fetch for other branches that might not be present locally
-  #suppressMessages(suppressWarnings(gert::git_fetch()))
-  # invisible(capture.output(gert::git_fetch()))
   gert::git_fetch()
 
   app <- shinyApp(
@@ -37,7 +36,7 @@ ghqc_notify_app <- function() {
     server = function(input, output, session) {
       ghqc_notify_server(
         id = "ghqc_notify_app",
-        milestone_list = milestone_list
+        open_milestone_names = open_milestone_names
       )
     }
   )