+++ R/app_03_record_server.R
@@ -7,7 +7,7 @@
 #' @importFrom rprojroot find_rstudio_root_file
 NULL
 
-ghqc_record_server <- function(id, all_milestones) {
+ghqc_record_server <- function(id, all_milestones_in, closed_milestones_in, all_milestone_names_in, all_closed_milestone_names_in) {
   moduleServer(id, function(input, output, session) {
     iv <- shinyvalidate::InputValidator$new()
     iv$add_rule("select_milestone", shinyvalidate::sv_required())
@@ -25,21 +25,32 @@ ghqc_record_server <- function(id, all_milestones) {
 
     ns <- session$ns
     report_trigger <- reactiveVal(FALSE)
+    all_milestones_rv <- reactiveVal(all_milestones_in)
+    closed_milestones_rv <- reactiveVal(closed_milestones_in)
+    all_milestone_names_rv <- reactiveVal(all_milestone_names_in)
+    closed_milestone_names_rv <- reactiveVal(all_closed_milestone_names_in)
+    issue_objects_rv <- reactiveVal(NULL)
+    statuses_rv <- reactiveVal(NULL)
+    selected_milestones_rv <- reactiveVal(NULL)
+    issues_with_unapproved_statuses_rv <- reactiveVal(NULL)
+    issues_with_open_checklists_rv <- reactiveVal(NULL)
 
     observe({
       waiter_hide()
     })
 
-    closed_milestones <- reactive({
+    reactive({
       w_gh <- create_waiter(ns, sprintf("Fetching Milestone data for %s in %s...", .le$repo, .le$org))
       w_gh$show()
       on.exit(w_gh$hide())
 
+      closed_milestones <- closed_milestones_rv()
+      req(closed_milestones)
+
       tryCatch(
         {
-          closed_milestones <- get_closed_milestone_names()
-          milestone_list_url <- get_milestone_list_url()
           if (length(closed_milestones) == 0) {
+            milestone_list_url <- get_milestone_list_url()
             warn_icon_html <- "<span style='font-size: 24px; vertical-align: middle;'>&#9888;</span>"
             showModal(
               modalDialog(
@@ -61,7 +72,6 @@ ghqc_record_server <- function(id, all_milestones) {
             )
             warn(.le$logger, glue::glue("There were no closed Milestones found in {.le$org}/{.le$repo}. Close relevant Milestones on GitHub to indicate finished QC."))
           } # length(closed_milestones) == 0
-          rev(closed_milestones)
         },
         error = function(e) {
           error(.le$logger, glue::glue("There was an error retrieving closed Milestones: {conditionMessage(e)}"))
@@ -72,26 +82,30 @@ ghqc_record_server <- function(id, all_milestones) {
 
 
     observeEvent(input$closed_only, {
+      closed_milestone_names <- closed_milestone_names_rv()
+      all_milestone_names <- all_milestone_names_rv()
+      req(closed_milestone_names, all_milestone_names)
+
       # if closed
       if (input$closed_only) {
-        placeholder <- ifelse(length(closed_milestones()) == 0, "No closed Milestones", "Select closed Milestones")
+        placeholder <- ifelse(length(closed_milestone_names) == 0, "No closed Milestones", "Select closed Milestones")
 
         updateSelectizeInput(
           session,
           "select_milestone",
-          choices = closed_milestones(),
+          choices = closed_milestone_names,
           options = list(placeholder = placeholder)
         )
       }
 
       # if not closed
       else {
-        placeholder <- ifelse(length(all_milestones) == 0, "No Milestones", "Select Milestones")
+        placeholder <- ifelse(length(all_milestone_names) == 0, "No Milestones", "Select Milestones")
 
         updateSelectizeInput(
           session,
           "select_milestone",
-          choices = all_milestones,
+          choices = all_milestone_names,
           options = list(placeholder = placeholder)
         )
       }
@@ -117,7 +131,19 @@ ghqc_record_server <- function(id, all_milestones) {
       w_check_status$show()
       on.exit(w_check_status$hide())
 
-      determine_modal_message_report(input$select_milestone)
+      all_milestones <- all_milestones_rv()
+      milestone_names <- input$select_milestone
+      milestone_objects <- purrr::map(milestone_names, ~ get_milestone_object_from_milestone_name(milestone_name = .x,
+                                                                                 milestone_objects = all_milestones
+                                                                                 ))
+      selected_milestones_rv(milestone_objects)
+
+      res <- determine_modal_message_report(milestone_objects)
+      issue_objects_rv(res$issue_objects)
+      statuses_rv(res$statuses)
+      issues_with_unapproved_statuses_rv(res$issues_with_unapproved_statuses)
+      issues_with_open_checklists_rv(res$issues_with_open_checklists)
+      return(res)
     })
 
     observeEvent(input$generate_report, {
@@ -131,14 +157,13 @@ ghqc_record_server <- function(id, all_milestones) {
               actionButton(ns("return"), "Return"),
               style = "text-align: right;"
             ),
-            HTML(paste(glue::glue("Upon completion of QC, It is recommended that:
+            HTML(paste(glue::glue("When QC is complete, it is recommended that:
             <ul>
-            <li>All selected Milestones are closed</li>
-            <li>All Issues within selected Milestones are closed</li>
-            <li>All {get_checklist_display_name_var()} items within relevant Issues are completed</li>
+            <li>Milestones are closed</li>
+            <li>{get_checklist_display_name_var(capitalized = TRUE, plural = TRUE)} are complete</li>
+            <li>Issues are approved with <code>ghqc_status_app()</code></li>
             </ul>
-            <br>
-            You may want to review the following items on GitHub for outstanding QC progress:<br><br>"),
+            You may want to review the following on GitHub for outstanding QC progress:<br>"),
                        modal_check()$message)),
             tags$style(HTML("
         .modal-content {
@@ -169,8 +194,12 @@ ghqc_record_server <- function(id, all_milestones) {
       on.exit(w_generate_report$hide())
 
       tryCatch({
-        pdf_path <- ghqc_report(
-          milestone_names = input$select_milestone,
+        pdf_path <- ghqc_record(
+          milestone_objects = selected_milestones_rv(),
+          issue_objects = issue_objects_rv(),
+          statuses = statuses_rv(),
+          issues_with_unapproved_statuses = issues_with_unapproved_statuses_rv(),
+          issues_with_open_checklists = issues_with_open_checklists_rv(),
           input_name = input$pdf_name,
           just_tables = input$just_tables,
           location = input$pdf_location
@@ -178,10 +207,14 @@ ghqc_record_server <- function(id, all_milestones) {
 
         showModal(
           modalDialog(
-            title = tags$div(modalButton("Dismiss"), style = "text-align: right;"),
+            title = tags$div(
+              tags$span("QC recorded", style = "float: left; font-weight: bold; font-size: 20px; margin-top: 5px;"),
+              modalButton("Dismiss"),
+              style = "text-align: right;"
+            ),
             footer = NULL,
             easyClose = TRUE,
-            tags$p(glue::glue("QC Record generated successfully: {pdf_path}"))
+            tags$p(glue::glue("PDF location: {pdf_path}"))
           )
         ) #showModal
       },
@@ -202,9 +235,21 @@ ghqc_record_server <- function(id, all_milestones) {
     })
 
     observeEvent(input$reset, {
+      session$reload()
       debug(.le$logger, glue::glue("App was reset through the reset button."))
       reset_triggered(TRUE)
-      session$reload()
+
+      all_milestones <- get_all_non_empty_ghqc_milestone_objects()
+      closed_milestones <- get_closed_milestone_objects_from_all_milestone_objects(all_milestones)
+      all_milestone_names <- get_milestone_names_from_milestone_objects(all_milestones)
+      closed_milestone_names <- get_milestone_names_from_milestone_objects(closed_milestones)
+
+      all_milestones_rv(all_milestones)
+      closed_milestones_rv(closed_milestones)
+      all_milestone_names_rv(all_milestone_names)
+      closed_milestone_names_rv(closed_milestone_names)
+
+
     })
 
     observeEvent(input$proceed, {